+++ IntegrityService/Jobs/IMonitor.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace IntegrityService.Jobs
+{
+    internal interface IMonitor : IDisposable
+    {
+        void Start();
+        void Stop();
+    }
+}
