+++ IntegrityService/Settings.cs
@@ -5,7 +5,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
-using IntegrityService.Utils;
+using IntegrityService.IO;
 
 namespace IntegrityService
 {
