--- IntegrityService/Utils/IMonitor.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace IntegrityService.Utils
-{
-    internal interface IMonitor: IDisposable
-    {
-        void Start();
-        void Stop();
-    }
-}
