+++ IntegrityService/Worker.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Threading;
 using System.Threading.Tasks;
+using IntegrityService.Data;
+using IntegrityService.Jobs;
 using IntegrityService.Utils;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -9,16 +11,16 @@ namespace IntegrityService
 {
     public class Worker : BackgroundService
     {
-        private readonly ILogger<Worker> _logger;
-        private readonly FileSystemMonitor _fsMonitor;
         private readonly BackgroundWorkerQueue _backgroundWorkerQueue;
-        private RegistryMonitor _regMonitor;
+        private readonly FileSystemMonitorJob _fsMonitor;
+        private readonly ILogger<Worker> _logger;
+        private RegistryMonitorJob _regMonitor;
 
         public Worker(ILogger<Worker> logger, BackgroundWorkerQueue backgroundWorkerQueue)
         {
             _logger = logger;
-            _fsMonitor = new FileSystemMonitor(_logger);
             _backgroundWorkerQueue = backgroundWorkerQueue;
+            _fsMonitor = new FileSystemMonitorJob(_logger);
             Database.Start();
         }
 
@@ -37,7 +39,7 @@ namespace IntegrityService
 
             if (Settings.Instance.EnableRegistryMonitoring)
             {
-                _regMonitor = new RegistryMonitor(_logger);
+                _regMonitor = new RegistryMonitorJob(_logger);
                 _regMonitor.Start();
             }
 
@@ -57,6 +59,40 @@ namespace IntegrityService
                 }
 #pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
                 await Task.Delay(Settings.Instance.HeartbeatInterval * 1000, stoppingToken);
+            }
+        }
+
+        private void Cleanup()
+        {
+            // Cleanup members here
+            _fsMonitor.Stop();
+            _fsMonitor.Dispose();
+
+            if (Settings.Instance.EnableRegistryMonitoring)
+            {
+                _regMonitor.Stop();
+                _regMonitor.Dispose();
+            }
+
+            Database.Stop();
+        }
+
+        private bool Handler(CtrlType signal)
+        {
+            switch (signal)
+            {
+                case CtrlType.CtrlBreakEvent:
+                case CtrlType.CtrlCEvent:
+                case CtrlType.CtrlLogoffEvent:
+                case CtrlType.CtrlShutdownEvent:
+                case CtrlType.CtrlCloseEvent:
+                    _logger.LogInformation("Worker stopped at: {time}", DateTimeOffset.Now);
+                    Cleanup();
+                    Environment.Exit(0);
+                    return false;
+
+                default:
+                    return false;
             }
         }
 
@@ -80,14 +116,14 @@ namespace IntegrityService
             try
             {
                 await Task.Run(() =>
-                       {
-                           _logger.LogInformation(
-                               "File discovery not completed. Initiating file system discovery. It will take time.");
-                           var fsDiscovery = new FileSystemDiscovery(_logger);
-                           fsDiscovery.Start();
-                           Settings.Instance.IsFileDiscoveryCompleted = true;
-                           _logger.LogInformation("File system discovery completed.");
-                       },
+                {
+                    _logger.LogInformation(
+                        "File discovery not completed. Initiating file system discovery. It will take time.");
+                    var fsDiscovery = new FileSystemDiscoveryJob(_logger);
+                    fsDiscovery.Start();
+                    Settings.Instance.IsFileDiscoveryCompleted = true;
+                    _logger.LogInformation("File system discovery completed.");
+                },
                        token);
             }
             catch (Exception ex)
@@ -95,40 +131,6 @@ namespace IntegrityService
                 tcs.SetException(ex);
             }
             return tcs.Task;
-        }
-
-        private bool Handler(CtrlType signal)
-        {
-            switch (signal)
-            {
-                case CtrlType.CtrlBreakEvent:
-                case CtrlType.CtrlCEvent:
-                case CtrlType.CtrlLogoffEvent:
-                case CtrlType.CtrlShutdownEvent:
-                case CtrlType.CtrlCloseEvent:
-                    _logger.LogInformation("Worker stopped at: {time}", DateTimeOffset.Now);
-                    Cleanup();
-                    Environment.Exit(0);
-                    return false;
-
-                default:
-                    return false;
-            }
-        }
-
-        private void Cleanup()
-        {
-            // Cleanup members here
-            _fsMonitor.Stop();
-            _fsMonitor.Dispose();
-
-            if (Settings.Instance.EnableRegistryMonitoring)
-            {
-                _regMonitor.Stop();
-                _regMonitor.Dispose();
-            }
-
-            Database.Stop();
         }
     }
 }
